requirements.txt
fastapi
uvicorn
Pillow
pydantic
python-multipart
